move category parsing to separate file
