+++ crates/html-bindgen/src/parse/elements/mod.rs
@@ -4,6 +4,9 @@ use convert_case::{Case, Casing};
 use serde::{Deserialize, Serialize};
 
 use super::{Attribute, AttributeType, ParsedCategory, ParsedRelationship};
+use categories::parse_categories;
+
+mod categories;
 
 /// The parsed values converted from the raw spec
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -198,64 +201,6 @@ fn parse_kinds(kind: String) -> String {
         other => panic!("unknown category: {other}"),
     };
     s.to_owned()
-}
-
-fn parse_categories(mut line: &str) -> Vec<String> {
-    if line.contains("Zero or more") {
-        let mut iter = line.split("Zero or more");
-        let _ = iter.next().unwrap();
-        line = iter.next().unwrap();
-        let mut output = vec![];
-        for word in line.split("and") {
-            output.push(parse_category(word));
-        }
-        output
-    } else if line.starts_with("Inside") {
-        line = line.strip_prefix("Inside").unwrap();
-        vec![parse_category(line)]
-    } else {
-        vec![parse_category(line)]
-    }
-}
-
-fn parse_category(mut line: &str) -> String {
-    if line.starts_with("If the") {
-        if line.contains(":") {
-            let mut iter = line.split(":");
-            _ = iter.next().unwrap();
-            line = iter.next().unwrap();
-        }
-    }
-
-    if line.starts_with("Where ") {
-        line = line.strip_prefix("Where ").unwrap();
-        if line.ends_with("is expected.") {
-            line = line.strip_suffix(" is expected.").unwrap();
-        } else if line.ends_with("are expected.") {
-            line = line.strip_suffix(" are expected.").unwrap();
-        };
-    }
-
-    if line.contains(",") {
-        let mut iter = line.split(",");
-        line = iter.next().unwrap();
-    }
-    if line.contains(".") {
-        let mut iter = line.split(".");
-        line = iter.next().unwrap();
-    }
-
-    if line.contains("content") {
-        let mut iter = line.split("content");
-        line = iter.next().unwrap();
-    }
-
-    if line.contains("elements") {
-        let mut iter = line.split("elements");
-        line = iter.next().unwrap();
-    }
-
-    line.trim().to_lowercase().to_owned()
 }
 
 fn parse_content_categories(categories: &[String]) -> Vec<ParsedCategory> {
